apple: Move xnu and apple_libc up a level in the new structure
There isn't any need to group these two modules.
