+++ src/new/apple_libc/mod.rs
@@ -1,0 +1,3 @@
+//! Entrypoint for Apple headers, usually found as part of the Xcode SDK.
+
+pub(crate) mod signal;
